requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
djangorestframework==3.12.4
ecdsa==0.14.1
idna==2.10
pyasn1==0.4.8
python-jose==3.2.0
python-keycloak==0.24.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
